storage: move DefaultDeclareKeys to batcheval
